+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * refactor: Consolidate boilerplate code for getting the `ObjectCache`.
 * refactor: Move transformer query logic to `TransformerRegistry`.
 * refactor: Mover more logic to `TransformerRegistry`.
+* refactor: Move `MainTransformer` to its own namespace.
 
 ## 0.5.4